Update code for environment branch

+++ Backend/src/controllers/authController.js
@@ -1,5 +1,8 @@
 // import bcrypt from "bcryptjs";
+// import fs from "fs";
+// import mongoose from "mongoose";
 // import User from "../models/User.js";
+// import { getGridFSBucket } from "../config/gridfs.js";
 
 // // SIGNUP
 // export const signupUser = async (req, res) => {
@@ -22,10 +25,35 @@
 
 //     const hashedPassword = await bcrypt.hash(password, 10);
 
+//     let profilePictureId = null;
+
+//     if (req.file) {
+//       const gfsBucket = getGridFSBucket();
+
+//       const uploadStream = gfsBucket.openUploadStream(req.file.originalname, {
+//         contentType: req.file.mimetype,
+//       });
+
+//       const readStream = fs.createReadStream(req.file.path);
+
+//       await new Promise((resolve, reject) => {
+//         readStream
+//           .pipe(uploadStream)
+//           .on("error", reject)
+//           .on("finish", resolve);
+//       });
+
+//       profilePictureId = uploadStream.id;
+
+//       // remove temporary local file after saving to GridFS
+//       fs.unlinkSync(req.file.path);
+//     }
+
 //     const newUser = await User.create({
 //       username,
 //       email,
 //       password: hashedPassword,
+//       profilePictureId,
 //     });
 
 //     res.status(201).json({
@@ -34,6 +62,10 @@
 //         id: newUser._id,
 //         username: newUser.username,
 //         email: newUser.email,
+//         profilePictureId: newUser.profilePictureId,
+//         profilePictureUrl: newUser.profilePictureId
+//           ? `http://localhost:5000/api/auth/profile-picture/${newUser.profilePictureId}`
+//           : "",
 //       },
 //     });
 //   } catch (error) {
@@ -70,8 +102,47 @@
 //         id: user._id,
 //         username: user.username,
 //         email: user.email,
+//         profilePictureId: user.profilePictureId,
+//         profilePictureUrl: user.profilePictureId
+//           ? `http://localhost:5000/api/auth/profile-picture/${user.profilePictureId}`
+//           : "",
 //       },
 //     });
+//   } catch (error) {
+//     res.status(500).json({ message: error.message });
+//   }
+// };
+
+// // STREAM PROFILE PICTURE FROM GRIDFS
+// export const getProfilePicture = async (req, res) => {
+//   try {
+//     const { id } = req.params;
+
+//     if (!mongoose.Types.ObjectId.isValid(id)) {
+//       return res.status(400).json({ message: "Invalid image id" });
+//     }
+
+//     const fileId = new mongoose.Types.ObjectId(id);
+//     const gfsBucket = getGridFSBucket();
+
+//     const files = await mongoose.connection.db
+//       .collection("profilePictures.files")
+//       .find({ _id: fileId })
+//       .toArray();
+
+//     if (!files || files.length === 0) {
+//       return res.status(404).json({ message: "Image not found" });
+//     }
+
+//     const file = files[0];
+//     res.set("Content-Type", file.contentType || "image/jpeg");
+
+//     const downloadStream = gfsBucket.openDownloadStream(fileId);
+//     downloadStream.on("error", () => {
+//       res.status(404).json({ message: "Error reading image" });
+//     });
+
+//     downloadStream.pipe(res);
 //   } catch (error) {
 //     res.status(500).json({ message: error.message });
 //   }
@@ -124,8 +195,9 @@ export const signupUser = async (req, res) => {
 
       profilePictureId = uploadStream.id;
 
-      // remove temporary local file after saving to GridFS
-      fs.unlinkSync(req.file.path);
+      if (fs.existsSync(req.file.path)) {
+        fs.unlinkSync(req.file.path);
+      }
     }
 
     const newUser = await User.create({
@@ -133,6 +205,10 @@ export const signupUser = async (req, res) => {
       email,
       password: hashedPassword,
       profilePictureId,
+      // these stay empty by default for now
+      aboutMe: "",
+      skills: [],
+      relationshipStatus: "",
     });
 
     res.status(201).json({
@@ -141,6 +217,9 @@ export const signupUser = async (req, res) => {
         id: newUser._id,
         username: newUser.username,
         email: newUser.email,
+        aboutMe: newUser.aboutMe,
+        skills: newUser.skills,
+        relationshipStatus: newUser.relationshipStatus,
         profilePictureId: newUser.profilePictureId,
         profilePictureUrl: newUser.profilePictureId
           ? `http://localhost:5000/api/auth/profile-picture/${newUser.profilePictureId}`
@@ -181,6 +260,9 @@ export const loginUser = async (req, res) => {
         id: user._id,
         username: user.username,
         email: user.email,
+        aboutMe: user.aboutMe,
+        skills: user.skills,
+        relationshipStatus: user.relationshipStatus,
         profilePictureId: user.profilePictureId,
         profilePictureUrl: user.profilePictureId
           ? `http://localhost:5000/api/auth/profile-picture/${user.profilePictureId}`
